requirements.txt
# TaxoCapsNet - Complete Requirements
# Core Deep Learning and ML
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Additional ML Models
xgboost==1.7.6
lightgbm==4.0.0

# Visualization and Analysis  
matplotlib==3.7.2
seaborn==0.12.2

# Interpretability (SHAP)
shap==0.42.1

# Flask API Server
flask==2.3.3
flask-cors==4.0.0

# Optional: ngrok for public tunneling
pyngrok==6.0.0

# Configuration and Utils
PyYAML==6.0.1
tqdm==4.66.1
joblib==1.3.2

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Testing (optional)
pytest==7.4.2
pytest-cov==4.1.0

# Development Tools (optional)
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# System utilities
psutil==5.9.5
requests==2.31.0